#11 - Move Logger to config package
